store: widen SetLocalPref to int64 for 32-bit builds
BGP local preference is a 32-bit unsigned value (0..4294967295), which
does not fit a 32-bit int. The bound check compared against the untyped
constant 4294967295, so `go build` for a 32-bit target (GOARCH=arm, 386)
failed with "overflows int" — caught by the linux/arm release target.
Widen the field to int64 (it holds the full range everywhere) and parse
the form value with ParseInt. Verified: linux/arm and linux/386 now build.

+++ internal/store/policies.go
@@ -56,7 +56,7 @@ type Policy struct {
 	OriginASSetID sql.NullInt64
 	// ROV is RPKI route-origin validation: off, log, or reject.
 	ROV          string
-	SetLocalPref int
+	SetLocalPref int64
 
 	// export
 	AnnounceEverything   bool
